remove Producer header, no need
